refactor: export ProgressItem
