+++ src/types/model.d.ts
@@ -1,0 +1,10 @@
+interface ProgressItem {
+    total: number;
+    loaded: number;
+    file: string;
+    progress: number;
+    name: string;
+    status: string;
+}
+
+export type { ProgressItem };
